refactor DynamicServicesSettings
